% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPivot.R
\name{getPivot}
\alias{getPivot}
\title{Get pivot element coordinates}
\usage{
getPivot(mat)
}
\arguments{
\item{input_matrix}{A matrix prepared for pivoting.}
}
\value{
Matrix of coordinates of possible pivots.
}
\description{
Identifies the pivot position.
}
\examples{
example_matrix <- matrix(c(4, 1, 8, 2, 3, 1, 0, 4, 3), nrow = 3, byrow = TRUE)
modified_matrix <- modifyMatrixForPivoting(example_matrix)
getPivot(modified_matrix)

}
