% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotingStep.R
\name{pivotingStep}
\alias{pivotingStep}
\title{Perform one pivoting step}
\usage{
pivotingStep(mat, x, y, D = 1)
}
\arguments{
\item{mat}{The current pivot matrix.}

\item{x}{The row index of the pivot element.}

\item{y}{The column index of the pivot element.}

\item{D}{A scalar used to calculate updated values.}
}
\value{
A list with the updated matrix and the pivot value used for the next step.
}
\description{
Executes one step of the pivoting algorithm.
}
\examples{
example_matrix <- matrix(c(4, 1, 8, 2, 3, 1, 0, 4, 3), nrow = 3, byrow = TRUE)
modified_matrix <- modifyMatrixForPivoting(example_matrix)
writeMatrixForPivoting(modified_matrix)
pivots <- getPivot(modified_matrix)
pivoted_matrix <- pivotingStep(modified_matrix, x=pivots[1,1], y=pivots[1,2], D=1)$mat
writeMatrixForPivoting(pivoted_matrix)

}
