% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDominatedStrategies.R
\name{removeDominatedStrategies}
\alias{removeDominatedStrategies}
\title{Remove dominated strategies from a payoff matrix}
\usage{
removeDominatedStrategies(
  mat,
  strictly = TRUE,
  granularity = 0.05,
  skip_combinations = FALSE
)
}
\arguments{
\item{mat}{A numeric payoff matrix.}

\item{strictly}{Logical. If \code{TRUE}, uses strict domination; if \code{FALSE}, uses weak domination.}

\item{granularity}{A numeric value between 0 and 1 indicating the step size for combinations of weights. Must be evenly divisible into 1.}

\item{skip_combinations}{Logical. If \code{TRUE}, only single strategies are checked for domination, not combinations.}
}
\value{
A reduced matrix with dominated strategies removed.
}
\description{
This function removes dominated rows and columns (strategies) from a matrix, using either strict or weak domination rules. It optionally checks for domination by combinations of strategies.
}
\examples{
example_matrix <- matrix(c(3,6,5,3,1,2,6,0,3), nrow = 3, byrow = TRUE)

# Without combinations
removeDominatedStrategies(example_matrix, strictly = TRUE, granularity = 0.1, skip_combinations = TRUE)

# With combinations and low granularity
removeDominatedStrategies(example_matrix, strictly = TRUE, granularity = 0.5, skip_combinations = FALSE)

# With combinations and high granularity
removeDominatedStrategies(example_matrix, strictly = TRUE, granularity = 0.1, skip_combinations = FALSE)

}
