% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveDiagonalMatrix.R
\name{solveDiagonalMatrix}
\alias{solveDiagonalMatrix}
\title{Solve a square diagonal matrix game}
\usage{
solveDiagonalMatrix(mat)
}
\arguments{
\item{mat}{A square diagonal matrix.}
}
\value{
A list containing:
\itemize{
\item \code{p}: Optimal mixed strategy for the row player.
\item \code{q}: Optimal mixed strategy for the column player.
\item \code{V}: Value of the game.
}
If there are multiple zeros on the diagonal, the function simplifies the matrix and recursively re-evaluates.
}
\description{
Analyzes a square diagonal matrix to determine optimal strategies and the value of a zero-sum game. Handles various cases including all-positive, all-negative, zero entries, and mixed-sign diagonals.
}
\examples{
example_matrix <- diag(c(2, 0, 4))
solveDiagonalMatrix(example_matrix)

example_matrix <- diag(c(-2, -3, 1, -4))
solveDiagonalMatrix(example_matrix)

example_matrix <- diag(c(-2, -3, -1, -4))
solveDiagonalMatrix(example_matrix)

example_matrix <- diag(c(-2, 0, -1, 0))
solveDiagonalMatrix(example_matrix)

}
