% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveMatrixByPivoting.R
\name{solveMatrixByPivoting}
\alias{solveMatrixByPivoting}
\title{Solve a matrix game using the pivoting algorithm}
\usage{
solveMatrixByPivoting(mat, steps = FALSE)
}
\arguments{
\item{mat}{A numeric matrix representing a two-player zero-sum game.}

\item{steps}{Logical. If \code{TRUE}, prints each pivoting step.}
}
\value{
A list containing:
\itemize{
\item \code{p}: Optimal mixed strategy for the row player.
\item \code{q}: Optimal mixed strategy for the column player.
\item \code{V}: Value of the game.
}
}
\description{
Uses a full pivoting method to solve a matrix game, returning optimal strategies and game value.
}
\examples{
# Without negative values
example_matrix <- matrix(c(4,1,8,2,3,1,0,4,3), nrow = 3, byrow = TRUE)
solveMatrixByPivoting(example_matrix, steps = TRUE)

# With negative values
example_matrix <- matrix(c(2,-1,6,0,1,-1,-2,2,1), nrow = 3, byrow = TRUE)
modified_matrix <- modifyMatrixForPivoting(example_matrix)
writeMatrixForPivoting(modified_matrix)
solveMatrixByPivoting(example_matrix, steps = TRUE)

}
