% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveNonsingularSquareMatrix.R
\name{solveNonsingularSquareMatrix}
\alias{solveNonsingularSquareMatrix}
\title{Solve a nonsingular square matrix}
\usage{
solveNonsingularSquareMatrix(mat)
}
\arguments{
\item{mat}{A square numeric matrix.}
}
\value{
A list containing:
\itemize{
\item \code{p}: Optimal mixed strategy for the row player.
\item \code{q}: Optimal mixed strategy for the column player.
\item \code{V}: Value of the game.
}
If the matrix is singular, the function adjusts it temporarily by adding 1 to all elements. If the matrix is not invertible or its rows are linearly dependent, an error is raised.
}
\description{
Solves a nonsingular square matrix by calculating a value \code{V}, a column strategy vector \code{q}, and a row strategy vector \code{p_t} using matrix inversion. If the determinant is zero, a temporary regularization is applied.
}
\examples{
example_matrix <- matrix(c(1,2,-1,2,-1,4,-1,4,-3), nrow = 3, byrow = TRUE)
example_matrix
solveNonsingularSquareMatrix(example_matrix)

# With determinant = 0
example_matrix <- matrix(c(0,1,-2,1,-2,3,-2,3,-4), nrow = 3, byrow = TRUE)
example_matrix
solveNonsingularSquareMatrix(example_matrix)

}
