% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMatrixForPivoting.R
\name{writeMatrixForPivoting}
\alias{writeMatrixForPivoting}
\title{Print matrix in pivoting format}
\usage{
writeMatrixForPivoting(mat)
}
\arguments{
\item{mat}{A matrix prepared for pivoting with labeling rows/columns.}
}
\description{
Nicely formats and prints a matrix used in a pivoting algorithm, showing variable labels and separators.
}
\examples{
example_matrix <- matrix(c(4, 1, 8, 2, 3, 1, 0, 4, 3), nrow = 3, byrow = TRUE)
example_matrix
modified_matrix <- modifyMatrixForPivoting(example_matrix)
writeMatrixForPivoting(modified_matrix)

}
